+++ app/src/main/java/app/urv/manager/ui/model/SupportedVersionInfo.kt
@@ -1,0 +1,6 @@
+package app.urv.manager.ui.model
+
+data class SupportedVersionInfo(
+    val version: String,
+    val experimental: Boolean = false
+)